requirements.txt
discord.py[voice]
python-dotenv
Daug==2023.12.9.1
